// packages/@aws-cdk/integ-runner/lib/workers/common.d.ts
export interface IntegTestWorkerConfig extends IntegTestInfo {
    /**
     * A list of any destructive changes
     *
     * @default []
     */
    readonly destructiveChanges?: DestructiveChange[];
}
/**
 * Information on any destructive changes
 */
export interface DestructiveChange {
    /**
     * The logicalId of the resource with a destructive change
     */
    readonly logicalId: string;
    /**
     * The name of the stack that contains the destructive change
     */
    readonly stackName: string;
    /**
     * The impact of the destructive change
     */
    readonly impact: ResourceImpact;
}
// ...
export interface SnapshotVerificationOptions {
    /**
     * Retain failed snapshot comparisons
     *
     * @default false
     */
    readonly retain?: boolean;
    /**
     * Verbose mode
     *
     * @default false
     */
    readonly verbose?: boolean;
}
// ...
export declare enum DiagnosticReason {
    /**
     * The integration test failed because there
     * is not existing snapshot
     */
    NO_SNAPSHOT = "NO_SNAPSHOT",
    /**
     * The integration test failed
     */
    TEST_FAILED = "TEST_FAILED",
    /**
     * There was an error running the integration test
     */
    TEST_ERROR = "TEST_ERROR",
    /**
     * The snapshot test failed because the actual
     * snapshot was different than the expected snapshot
     */
    SNAPSHOT_FAILED = "SNAPSHOT_FAILED",
    /**
     * The snapshot test failed because there was an error executing it
     */
    SNAPSHOT_ERROR = "SNAPSHOT_ERROR",
    /**
     * The snapshot test succeeded
     */
    SNAPSHOT_SUCCESS = "SNAPSHOT_SUCCESS",
    /**
     * The integration test succeeded
     */
    TEST_SUCCESS = "TEST_SUCCESS",
    /**
     * The assertion failed
     */
    ASSERTION_FAILED = "ASSERTION_FAILED"
}
/**
 * Integration test diagnostics
 * This is used to report back the status of each test
 */
export interface Diagnostic {
    /**
     * The name of the test
     */
    readonly testName: string;
    /**
     * The name of the stack
     */
    readonly stackName: string;
    /**
     * The diagnostic message
     */
    readonly message: string;
    /**
     * The time it took to run the test
     */
    readonly duration?: number;
    /**
     * The reason for the diagnostic
     */
    readonly reason: DiagnosticReason;
    /**
     * Additional messages to print
     */
    readonly additionalMessages?: string[];
    /**
     * Relevant config options that were used for the integ test
     */
    readonly config?: Record<string, any>;
}